dom/security/test/csp/test_evalscript_blocked_by_strict_dynamic.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <meta http-equiv="Content-Security-Policy" 
        content="script-src 'nonce-foobar' 'strict-dynamic'">
  <title>Bug 1439330  - CSP: eval is not blocked if 'strict-dynamic' is enabled
  </title>
  <script nonce="foobar" type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<script nonce="foobar">

/* Description of the test:
 * We apply the script-src 'nonce-foobar' 'strict-dynamic' CSP and
 * check if the eval function is blocked correctly by the CSP.
 */

SimpleTest.waitForExplicitFinish();

// start the test
try {
  eval("1");
  ok(false, "eval should be blocked by CSP");
}
catch (ex) {
  ok(true, "eval blocked by CSP");
}

SimpleTest.finish();

</script>
</body>
</html>
